move row and cell into 'components' directory
