+++ packages/nimble-components/src/table/components/cell/index.ts
@@ -1,6 +1,6 @@
 import { observable } from '@microsoft/fast-element';
 import { DesignSystem, FoundationElement } from '@microsoft/fast-foundation';
-import type { TableFieldValue } from '../types';
+import type { TableFieldValue } from '../../types';
 import { styles } from './styles';
 import { template } from './template';
