+++ packages/nimble-components/src/table/components/cell/template.ts
