+++ packages/nimble-components/src/table/components/row/styles.ts
