+++ packages/nimble-components/src/table/components/row/template.ts
